LayoutTests/imported/w3c/web-platform-tests/css/css-font-loading/nonexistent-file-url-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
CONSOLE MESSAGE: Not allowed to load local resource: thisisnotarealfile.otf
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame

PASS nonexistent-file-url
